+++ Sources/PulseBar/Models/SystemMetrics.swift
@@ -8,13 +8,17 @@ enum MemoryPressure: Equatable {
 
 struct SystemMetrics: Equatable {
     var cpuPercent: Double = 0
+    var deviceTemperatureCelsius: Double?
     var memoryUsedBytes: UInt64 = 0
     var memoryTotalBytes: UInt64 = 0
+    var appMemoryBytes: UInt64 = 0
     var inactiveBytes: UInt64 = 0
     var purgeableBytes: UInt64 = 0
+    var externalBytes: UInt64 = 0
     var compressorBytes: UInt64 = 0
     var wiredBytes: UInt64 = 0
     var freeBytes: UInt64 = 0
+    var swapUsedBytes: UInt64 = 0
     var pressure: MemoryPressure = .normal
 
     var memoryUsedRatio: Double {
@@ -24,5 +28,9 @@ struct SystemMetrics: Equatable {
 
     var reclaimableBytes: UInt64 {
         inactiveBytes + purgeableBytes
+    }
+
+    var cachedFilesBytes: UInt64 {
+        externalBytes + purgeableBytes
     }
 }